requirements.txt
requests~=2.28.2
PyYAML~=6.0